Code Used for reportmaking

+++ src/Individual.java
@@ -3,18 +3,18 @@
  */
 public class Individual {
     Genome genome;
-    int fitness;
+    double fitness;
     double proportionate_fitness_start;
     double proportionate_fitness;
 
-    public Individual(Genome genome, int fitness, double proportionate_fitness, double proportionate_fitness_start){
+    public Individual(Genome genome, double fitness, double proportionate_fitness, double proportionate_fitness_start){
         this.genome = genome;
         this.proportionate_fitness = proportionate_fitness;
         this.fitness = fitness;
         this.proportionate_fitness_start = proportionate_fitness_start;
     }
 
-    public Individual(Genome genome, int fitness){
+    public Individual(Genome genome, double fitness){
         this.genome = genome;
         this.fitness = fitness;
     }
@@ -39,7 +39,7 @@ public class Individual {
         return proportionate_fitness;
     }
 
-    public int getFitness(){
+    public double getFitness(){
         return fitness;
     }
 }
